moved service specs from app/tests to /test, as issue #20 says
